Clean up code and move callback impl of IUnknown into CallbackBase.
